+++ datanode/service/market.go
@@ -25,7 +25,7 @@ type MarketStore interface {
 	GetByID(ctx context.Context, marketID string) (entities.Market, error)
 	GetByTxHash(ctx context.Context, txHash entities.TxHash) ([]entities.Market, error)
 	GetAllPaged(ctx context.Context, marketID string, pagination entities.CursorPagination, includeSettled bool) ([]entities.Market, entities.PageInfo, error)
-	ListSuccessorMarkets(ctx context.Context, marketID string, fullHistory bool) ([]entities.Market, error)
+	ListSuccessorMarkets(ctx context.Context, marketID string, fullHistory bool, pagination entities.CursorPagination) ([]entities.SuccessorMarket, entities.PageInfo, error)
 }
 
 type Markets struct {
@@ -89,6 +89,6 @@ func (m *Markets) GetAllPaged(ctx context.Context, marketID string, pagination e
 	return m.store.GetAllPaged(ctx, marketID, pagination, includeSettled)
 }
 
-func (m *Markets) ListSuccessorMarkets(ctx context.Context, marketID string, childrenOnly bool) ([]entities.Market, error) {
-	return m.store.ListSuccessorMarkets(ctx, marketID, childrenOnly)
+func (m *Markets) ListSuccessorMarkets(ctx context.Context, marketID string, childrenOnly bool, pagination entities.CursorPagination) ([]entities.SuccessorMarket, entities.PageInfo, error) {
+	return m.store.ListSuccessorMarkets(ctx, marketID, childrenOnly, pagination)
 }